requirements.txt
fastapi==0.122.0
numpy==1.26.4
opencv-python==4.7.0.72
pydantic==2.12.5
pyserial==3.5
requests==2.32.5
ultralytics==8.1.0
